GH24992 Admin Logger: Moved event class to internal API module
So that it can be used in any module.
Added documentation on writing listeners.
