--- nucleus/admin/rest/rest-service/src/main/java/org/glassfish/admin/rest/events/CommandInvokedEvent.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-package org.glassfish.admin.rest.events;
-
-import javax.security.auth.Subject;
-import org.glassfish.api.admin.ParameterMap;
-
-/**
- *
- * @author Ondro Mihalyi
- */
-public class CommandInvokedEvent {
-
-    public CommandInvokedEvent(String commandName, ParameterMap parameters, Subject subject) {
-        this.commandName = commandName;
-        this.parameters = parameters;
-        this.subject = subject;
-    }
-
-    private String commandName;
-    private ParameterMap parameters;
-    private Subject subject;
-
-    public String getCommandName() {
-        return commandName;
-    }
-
-    public ParameterMap getParameters() {
-        return parameters;
-    }
-
-    public Subject getSubject() {
-        return subject;
-    }
-
-
-
-}
