+++ nucleus/common/internal-api/src/main/java/org/glassfish/internal/api/events/InvokeEventService.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.admin.rest.events;
+package org.glassfish.internal.api.events;
 
 import jakarta.inject.Inject;
 import org.glassfish.hk2.api.messaging.Topic;
@@ -21,6 +21,12 @@ import org.glassfish.internal.api.Globals;
 import org.jvnet.hk2.annotations.Service;
 
 /**
+ * <p>
+ * A service to retrieve the correct HK2 topic to invoke global events.
+ * </p>
+ * <p>
+ * Do not inject this service. Instead, retrieve it using {@link InvokeEventService#get()}, which returns a service registered with the global service locator. This is needed to trigger events that can be received by any subscriber registered globally.
+ * </p>
  *
  * @author Ondro Mihalyi
  */
